+++ peer/roundrobin/list.go
@@ -24,7 +24,8 @@ import (
 	"time"
 
 	"go.uber.org/yarpc/api/peer"
-	"go.uber.org/yarpc/peer/peerlist/v2"
+	"go.uber.org/yarpc/peer/abstractlist"
+	"go.uber.org/zap"
 )
 
 type listConfig struct {
@@ -32,6 +33,7 @@ type listConfig struct {
 	shuffle  bool
 	failFast bool
 	seed     int64
+	logger   *zap.Logger
 }
 
 var defaultListConfig = listConfig{
@@ -64,6 +66,13 @@ func FailFast() ListOption {
 	}
 }
 
+// Logger specifies a logger.
+func Logger(logger *zap.Logger) ListOption {
+	return func(c *listConfig) {
+		c.logger = logger
+	}
+}
+
 // New creates a new round robin peer list.
 func New(transport peer.Transport, opts ...ListOption) *List {
 	cfg := defaultListConfig
@@ -71,19 +80,22 @@ func New(transport peer.Transport, opts ...ListOption) *List {
 		o(&cfg)
 	}
 
-	plOpts := []peerlist.ListOption{
-		peerlist.Capacity(cfg.capacity),
-		peerlist.Seed(cfg.seed),
+	plOpts := []abstractlist.Option{
+		abstractlist.Capacity(cfg.capacity),
+		abstractlist.Seed(cfg.seed),
+	}
+	if cfg.logger != nil {
+		plOpts = append(plOpts, abstractlist.Logger(cfg.logger))
 	}
 	if !cfg.shuffle {
-		plOpts = append(plOpts, peerlist.NoShuffle())
+		plOpts = append(plOpts, abstractlist.NoShuffle())
 	}
 	if cfg.failFast {
-		plOpts = append(plOpts, peerlist.FailFast())
+		plOpts = append(plOpts, abstractlist.FailFast())
 	}
 
 	return &List{
-		List: peerlist.New(
+		List: abstractlist.New(
 			"round-robin",
 			transport,
 			newPeerRing(),
@@ -94,5 +106,5 @@ func New(transport peer.Transport, opts ...ListOption) *List {
 
 // List is a PeerList which rotates which peers are to be selected in a circle
 type List struct {
-	*peerlist.List
+	*abstractlist.List
 }