+++ transport/http/close_idle_pre_go112.go
@@ -18,7 +18,10 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package yarpc // import "go.uber.org/yarpc"
+// +build !go1.12
 
-// Version is the current version of YARPC.
-const Version = "1.42.1"
+package http
+
+import "net/http"
+
+func closeIdleConnections(client *http.Client) {}