+++ transport/http/transport.go
@@ -306,6 +306,7 @@ func (a *Transport) Start() error {
 // Stop stops the HTTP transport.
 func (a *Transport) Stop() error {
 	return a.once.Stop(func() error {
+		closeIdleConnections(a.client)
 		a.connectorsGroup.Wait()
 		return nil
 	})