+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.42.1] - 2019-11-27 (Gobble)
+### Fixed
+- Simplified the flow of status change notifications for the gRPC and TChannel
+  transports to reduce the liklihood of deadlocks.
+- Increase default HTTP timeout to avoid stop timeout errors when the server
+  has a new idle connection.
+- Close idle connections when the transport is closed.
+
 ## [1.42.0] - 2019-10-31 (Spooky)
 ### Added
 - Added fail-fast option to peer lists.  With this option enabled, a peer list
@@ -1136,6 +1144,7 @@ This release requires regeneration of ThriftRW code.
 
 - Initial release.
 
+[1.42.1]: https://github.com/yarpc/yarpc-go/compare/v1.42.0...v1.42.1
 [1.42.0]: https://github.com/yarpc/yarpc-go/compare/v1.41.0...v1.42.0
 [1.41.0]: https://github.com/yarpc/yarpc-go/compare/v1.40.0...v1.41.0
 [1.40.0]: https://github.com/yarpc/yarpc-go/compare/v1.39.0...v1.40.0